modify app.rb
