+++ movie.rb
@@ -2,9 +2,10 @@ require_relative 'item'
 
 class Movie < Item
   attr_accessor :silent, :name
+  attr_reader :id, :publish_date, :archived
 
-  def initialize(name, silent, publish_date, id = nil)
-    super(id, publish_date)
+  def initialize(id, name, silent, publish_date, archived)
+    super(id, publish_date, archived)
     @silent = silent
     @name = name
   end
